resolved merge conflicts. Moved current <App> code to separate module
